adding option to align non-host reads to nt database for single-end RNA_Seq

+++ mRNA.nonHost_seq.sh
@@ -1,10 +1,12 @@
 #!/bin/bash
 
 # read arguments
-while getopts d: option
+pairEnd=false
+while getopts d:p option
 do
     case "$option" in
 	d) dirData=$OPTARG;;
+	p) pairEnd=true;;
     esac
 done
 
@@ -15,16 +17,28 @@ then
     echo -ne "$currentDate: aligning with kraken2..."
     sampleID=$(find $dirData -name "*.out.mate1")
     sampleID=$(echo $sampleID | sed -r 's/.out.mate1//g')
-    /mnt/rstor/SOM_PATH_RXS745U/bin/kraken2_2.0.8/kraken2 \
-	--db /mnt/rstor/SOM_PATH_RXS745U/genome/kraken \
-	--paired \
-	${sampleID}.out.mate1 \
-	${sampleID}.out.mate2 \
-	--output $sampleID.kraken2.out \
-	--report $sampleID.kraken2.report \
-	--report-zero-counts
-    # clean up
-    rm ${sampleID}.out.mate1
-    rm ${sampleID}.out.mate2
+    if $pairEnd
+    then
+	/mnt/rstor/SOM_PATH_RXS745U/bin/kraken2_2.0.8/kraken2 \
+	    --db /mnt/rstor/SOM_PATH_RXS745U/genome/kraken \
+	    --paired \
+	    ${sampleID}.out.mate1 \
+	    ${sampleID}.out.mate2 \
+	    --output $sampleID.kraken2.out \
+	    --report $sampleID.kraken2.report \
+	    --report-zero-counts
+	# clean up
+	rm ${sampleID}.out.mate1
+	rm ${sampleID}.out.mate2
+    else
+	/mnt/rstor/SOM_PATH_RXS745U/bin/kraken2_2.0.8/kraken2 \
+            --db /mnt/rstor/SOM_PATH_RXS745U/genome/kraken \
+            ${sampleID}.out.mate1 \
+            --output $sampleID.kraken2.out \
+            --report $sampleID.kraken2.report \
+            --report-zero-counts
+        # clean up
+        rm ${sampleID}.out.mate1
+    fi
     echo "done"
 fi
